refactor: Remove unused code and optimize rendering logic
